.history/news-mosaic/backend/requirements_20260206232212.txt
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv
httpx
numpy
scikit-learn
google-generativeai
aiohttp